# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "eth-events API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/event/list": {
            "post": {
                "description": "查询事件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "查询事件",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.EventListRo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bean.Event"
                            }
                        }
                    }
                }
            }
        },
        "/v1/task/add": {
            "post": {
                "description": "添加任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TASK"
                ],
                "summary": "添加任务",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.TaskAddRo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/v1/task/delete": {
            "post": {
                "description": "删除任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TASK"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.TaskDeleteRo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/bean.Resp"
                        }
                    }
                }
            }
        },
        "/v1/task/list": {
            "get": {
                "description": "查询任务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TASK"
                ],
                "summary": "查询任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Task"
                            }
                        }
                    }
                }
            }
        },
        "/v1/task/pause": {
            "post": {
                "description": "任务暂停与运行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TASK"
                ],
                "summary": "任务暂停与运行",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.TaskPauseRo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/bean.Resp"
                        }
                    }
                }
            }
        },
        "/v1/task/update": {
            "post": {
                "description": "更新任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TASK"
                ],
                "summary": "更新任务",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bean.TaskUpdateRo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/bean.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bean.BlockRo": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
        "bean.Event": {
            "type": "object",
            "additionalProperties": true
        },
        "bean.EventListRo": {
            "type": "object",
            "required": [
                "event",
                "taskId"
            ],
            "properties": {
                "blockRo": {
                    "$ref": "#/definitions/bean.BlockRo"
                },
                "cols": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "event": {
                    "type": "string"
                },
                "orderRo": {
                    "$ref": "#/definitions/bean.OrderRo"
                },
                "pageRo": {
                    "$ref": "#/definitions/bean.PageRo"
                },
                "taskId": {
                    "type": "integer"
                },
                "timeRo": {
                    "$ref": "#/definitions/bean.TimeRo"
                },
                "txRo": {
                    "$ref": "#/definitions/bean.TxRo"
                },
                "wheres": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/bean.Where"
                        }
                    }
                }
            }
        },
        "bean.OrderRo": {
            "type": "object",
            "properties": {
                "feilds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "orderType": {
                    "type": "string",
                    "enum": [
                        "ASC",
                        "DESC"
                    ]
                }
            }
        },
        "bean.PageRo": {
            "type": "object",
            "required": [
                "limit"
            ],
            "properties": {
                "cursor": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                }
            }
        },
        "bean.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "bean.TaskAddRo": {
            "type": "object",
            "required": [
                "abi",
                "chainId",
                "contract",
                "interval",
                "rpc",
                "start"
            ],
            "properties": {
                "abi": {
                    "type": "string"
                },
                "chainId": {
                    "type": "integer"
                },
                "contract": {
                    "type": "string"
                },
                "interval": {
                    "description": "轮询间隔，建议为区块出块间隔,例如：ethereum 12sec,BSC 3sec",
                    "type": "integer"
                },
                "rpc": {
                    "type": "string"
                },
                "start": {
                    "description": "开始区块，建议是合约部署时区块号",
                    "type": "integer"
                }
            }
        },
        "bean.TaskDeleteRo": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "bean.TaskPauseRo": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "pause": {
                    "description": "0-默认 1-暂停",
                    "type": "integer"
                }
            }
        },
        "bean.TaskUpdateRo": {
            "type": "object",
            "required": [
                "id",
                "rpc"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "rpc": {
                    "type": "string"
                }
            }
        },
        "bean.TimeRo": {
            "type": "object",
            "properties": {
                "begin": {
                    "type": "integer"
                },
                "end": {
                    "type": "integer"
                }
            }
        },
        "bean.TxRo": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                }
            }
        },
        "bean.Where": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "op": {
                    "description": "can be =、\u003e、\u003c、\u003c\u003e and any operator supported by sql-database",
                    "type": "string",
                    "default": "="
                },
                "value": {}
            }
        },
        "model.Task": {
            "type": "object",
            "properties": {
                "abi": {
                    "type": "string"
                },
                "chainId": {
                    "type": "integer"
                },
                "contract": {
                    "type": "string"
                },
                "current": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "interval": {
                    "description": "区块轮询间隔",
                    "type": "integer"
                },
                "paused": {
                    "description": "是否暂停",
                    "type": "integer"
                },
                "rpc": {
                    "type": "string"
                },
                "start": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        }
    }
}